+++ packages/time/src/init.lua
@@ -1,3 +1,4 @@
+local RefreshScheduler = require('./RefreshScheduler')
 local debounce = require('./debounce')
 local loopUntil = require('./loopUntil')
 local loopWhile = require('./loopWhile')
@@ -5,9 +6,12 @@ local noYield = require('./noYield')
 local throttle = require('./throttle')
 local yieldUntil = require('./yieldUntil')
 
+export type RefreshScheduler<Request, Result> = RefreshScheduler.RefreshScheduler<Request, Result>
+export type RefreshSchedulerOptions<Request> = RefreshScheduler.RefreshSchedulerOptions<Request>
 export type Interval = loopUntil.Interval
 
 return {
+    RefreshScheduler = RefreshScheduler,
     debounce = debounce,
     loopUntil = loopUntil,
     loopWhile = loopWhile,